out_modified.php
As of now the images are not being saved as a physical file instead the data is being transferred in the form of binary. If you look in to the out.php file, you will find these line :

foreach($_FILES as $file) {
$imgData = base64_encode(file_get_contents($file['tmp_name']));
$src = 'data: '.mime_content_type($img_file).';base64,'.$imgData;

You will have to replace the second line as the following, if not you will not be able to get the mime_content_type and as a result the image binary string gets corrupted. You may try and find it for yourself.

$src = 'data: '.mime_content_type($file['tmp_name']).';base64,'.$imgData;

The next part is saving the data or image by decoding the base64 into an upload directory. You will have to add these codes to implement it.

//Exploding and decoding the image
list($type, $src) = explode(';', $src);
list(, $src) = explode(',', $src);
$src = base64_decode($src);
//Putting the image in the parent directory
file_put_contents('image.png', $src);

One more thing, if you are using xampp on mac then you will have to issue read write properties to the folder.

I have attached a sample php file which you can replace to get the result.

Hope this would solve you problem :+1: